To preserve sanity, move things into a _descriptor.h for semblance of structure
